withMonitoredClient
